third_party/blink/web_tests/platform/generic/fast/forms/radio/radio-default-value-expected.txt

PASS el.value is "on"
PASS el.setAttribute("value", "foo"); el.value is "foo"
PASS el.checked = true; el.value is "foo"
PASS el.removeAttribute("value"); el.value is "on"
PASS el.value = "foo"; el.getAttribute("value") is "foo"
PASS successfullyParsed is true

TEST COMPLETE
